+++ src/Entity/Worklog.php
@@ -3,6 +3,7 @@
 namespace App\Entity;
 
 use App\Entity\Trait\DataProviderTrait;
+use App\Entity\Trait\SynchronizedEntityTrait;
 use App\Enum\BillableKindsEnum;
 use App\Repository\WorklogRepository;
 use Doctrine\DBAL\Types\Types;
@@ -17,7 +18,9 @@ use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
 class Worklog extends AbstractBaseEntity
 {
     use DataProviderTrait;
+    use SynchronizedEntityTrait;
 
+    // TODO: Rename to projectTrackerId.
     #[ORM\Column]
     private ?int $worklogId = null;