+++ src/Exception/NotAcceptableException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App\Exception;
+
+class NotAcceptableException extends \Exception
+{
+}