+++ src/Message/LeantimeDeleteMessage.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Message;
+
+readonly class LeantimeDeleteMessage
+{
+    public function __construct(
+        public int $dataProviderId,
+        public bool $asyncJobQueue,
+        public ?\DateTimeInterface $deletedAfter,
+    ) {
+    }
+}