+++ src/Service/HourReportService.php
@@ -59,7 +59,9 @@ class HourReportService
 
             $projectTicket->timesheets->add($timesheets);
 
-            $issueEpicName = $issue->getEpicName() ?? '';
+            $epics = $issue->getEpics();
+            // There should only be one tag per issue, but in case of multiple, we comma separate them in the list.
+            $issueEpicName = $epics->isEmpty() ? '' : implode(', ', $epics->map(fn ($epic) => $epic->getTitle())->toArray());
 
             if ($hourReportData->projectTags->containsKey($issueEpicName)) {
                 $projectTag = $hourReportData->projectTags->get((string) $issueEpicName);